--- src/Components/GoogleButton.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import "./Login.css"; 
-
-const GoogleButton = () => {
-  return (
-    <button className="google-button">
-      <img src="/google.png" alt="Sign in with Google" className="google-image" />
-    </button>
-  );
-};
-
-export default GoogleButton;
