--- src/Components/Login.css
@@ -1,81 +1,0 @@
-@font-face {
-  font-family: 'SeoulNamsanM';
-  src: url('https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_two@1.0/SeoulNamsanM.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  width: 100vw;
-  height: 100vh;
-  background-color: #fff;
-}
-
-.title-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 270px;
-  width: 427px;
-  height: 96px;
-}
-
-.subtitle{
-  font-size: 1rem;
-  font-weight: 340; 
-  color: #00462A;
-  margin: 0;
-  padding: 0;
-  line-height: 1;
-}
-
-.title {
-  font-family: 'SeoulNamsanM', sans-serif;
-  font-weight: 730;
-  font-size: 2.15rem;
-  color: #00462A;
-  margin: 5px;
-  padding: 0;
-  line-height: 1;
-  margin-bottom: 24px;
-}
-
-.description {
-  position: absolute;
-  top: 337px;
-  width: 348px;
-  text-align: center;
-  font-family: "Pretendard", sans-serif;
-  font-weight: 500;
-  font-size: 16.8px;
-  line-height: 21px;
-  color: #333333;
-}
-
-
-.google-button {
-  position: absolute;
-  top: 411px;
-  width: 249px;
-  height: 60px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-}
-
-.google-image {
-  width: 74%;
-  height: 74%;
-  object-fit: cover;
-}
