+++ src/Components/Login/GoogleButton.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "./LoginStyle.css"; 
+
+const GoogleButton = () => {
+  return (
+    <Button>
+      <Image src="/google.png" alt="Sign in with Google" />
+    </Button>
+  );
+};
+
+export default GoogleButton;
